Reset status of token when aes key is deleted
